+++ lib/wsv/status.rb
@@ -8,7 +8,10 @@ module Wsv
       400 => "Bad Request",
       403 => "Forbidden",
       404 => "Not Found",
-      405 => "Method Not Allowed"
+      405 => "Method Not Allowed",
+      408 => "Request Timeout",
+      414 => "URI Too Long",
+      431 => "Request Header Fields Too Large"
     }.freeze
 
     def self.reason(code)